+++ Changelog.md
@@ -1,6 +1,13 @@
 # Changelog
 
 ## [unreleased]
+- Allow inactive groups in the graders table to be toggled for display (#6778)
+- Enable plotly rendering for jupyter notebooks (#6871)
+
+## [v2.4.1]
+- Internal changes only
+
+## [v2.4.0]
 - Display error message on file upload failure (#6703)
 - Remark request grading: differentiate between old and new annotations using colour and remark flag (#6704)
 - Display detailed messages about grace period deductions for an assignment on student interface (#6706)